backend/requirements.txt
Flask==2.3.3
requests==2.31.0
python-snappy==0.6.1
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
ssz==0.2.4
pytest==7.4.3
pytest-mock==3.12.0
requests-mock==1.11.0 